+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Changelog
 
-### 16.1 (12.11.2019)
+### 0.16.9 (06.12.2019)
+
+* Added prop "textures" to component GLTF for overriding textures. 
+
+### 0.16.2 - 0.16.6 (05.12.2019)
+
+* Added component Screenshot.
+
+### 0.16.1 (12.11.2019)
 
 * Added component Sphere.
 
-### 16.0 (08.11.2019)
+### 0.16.0 (08.11.2019)
 
 * First public release.